feat!: Consume service account file as Path rather than String
Since upstream uses a `String` to represent a path, using
`ServiceAccount::load_from_file` means that non-UTF8 paths cannot be
used.

This is very awkward, and as we *also* use a `String` to represent the
path, this pushes error handling for this downstream, which means
writing ugly code to uphold a pointless restriction.

So we handle file reading ourselves.